add fields param to retrieve()

+++ ads_api/api/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Ad(models.Model):
     title = models.CharField(max_length=200)
     description = models.CharField(max_length=1000)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=8, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add=True)
 
     class Meta:
